demo/requirements.txt
requests
djangorestframework